+++ tedi/components/form/index.ts
@@ -10,6 +10,7 @@ export * from "./date-field";
 export * from "./date-picker/date-picker.component";
 export * from "./feedback-text/feedback-text.component";
 export * from "./label/label.component";
+export * from "./label-row";
 export * from "./number-field/number-field.component";
 export * from "./select";
 export * from "./slider/slider.component";